Mengubah tipe data variabel lat dan lon dari Float ke Double serta mengubah value dari listStory dari listOf() menjadi emptyList()

+++ app/src/main/java/com/example/dicodingstory/data/remote/response/StoryDetailResponse.kt
@@ -32,8 +32,8 @@ data class Story(
 	val createdAt: String? = null,
 
 	@field:SerializedName("lat")
-	val lat: Float? = null,
+	val lat: Double? = null,
 
 	@field:SerializedName("lon")
-	val lon: Float? = null
+	val lon: Double? = null
 )
